automation/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/requirements.in
#
attrs==24.2.0
    # via
    #   outcome
    #   trio
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via
    #   requests
    #   selenium
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
h11==0.14.0
    # via wsproto
idna==3.10
    # via
    #   requests
    #   trio
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via pytest-html
mako==1.3.5
    # via pytest-bdd
markupsafe==3.0.1
    # via
    #   jinja2
    #   mako
outcome==1.3.0.post0
    # via trio
packaging==24.1
    # via
    #   build
    #   pytest
    #   pytest-bdd
    #   webdriver-manager
parse==1.20.2
    # via
    #   parse-type
    #   pytest-bdd
parse-type==0.6.4
    # via pytest-bdd
pip-tools==7.4.1
    # via -r requirements/requirements.in
pluggy==1.5.0
    # via pytest
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pysocks==1.7.1
    # via urllib3
pytest==8.3.3
    # via
    #   -r requirements/requirements.in
    #   pytest-asyncio
    #   pytest-bdd
    #   pytest-html
    #   pytest-metadata
pytest-asyncio==0.24.0
    # via -r requirements/requirements.in
pytest-bdd==7.3.0
    # via -r requirements/requirements.in
pytest-html==4.1.1
    # via -r requirements/requirements.in
pytest-metadata==3.1.1
    # via pytest-html
python-dotenv==1.0.1
    # via
    #   -r requirements/requirements.in
    #   webdriver-manager
requests==2.32.3
    # via webdriver-manager
ruff==0.6.9
    # via -r requirements/requirements.in
selenium==4.25.0
    # via -r requirements/requirements.in
six==1.16.0
    # via parse-type
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.26.2
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   pytest-bdd
    #   selenium
urllib3[socks]==2.2.3
    # via
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via -r requirements/requirements.in
websocket-client==1.8.0
    # via selenium
wheel==0.44.0
    # via pip-tools
wsproto==1.2.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
